Dettach electron process on Windows
On Windows 10, Node 6.3.0, Chromium 52.0.2743.82, Electron 1.3.1 the spawned process is attached unless "detached: true" is passed to proc.spawn().

On Windows, this causes the newly spawned electron process to be considered attached and killed along with the parent when app.quit() is called. This fixed for me on Windows. I couldn't test it on Linux or OS X.